cmd: move the cmd package inside a subdir named mapt
now when go install is used it creates the binary with name mapt
earlier the created binary was named cmd
