--- cmd/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/redhat-developer/mapt/cmd/cmd"
-
-func main() {
-	cmd.Execute()
-}
