+++ cmd/mapt/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/redhat-developer/mapt/cmd/mapt/cmd"
+
+func main() {
+	cmd.Execute()
+}
